refactor(cli): flatten computer-use command module
Move the Computer Use command slice to the neutral command root and update its internal export without changing public CLI or compatibility behavior.
