+++ turbo/apps/cli/src/commands/computer-use/output-artifacts.ts
